continued flailing in an ocean of confusion
