+++ docs/index.html
@@ -10,34 +10,23 @@
     <div class="wrapper">
         <h1>BLACKJACK!</h1>
         <main>
-            <textarea class="instructional-text-area" id="instructionsText" width=>Click Start to Play!</textarea>
+            <!-- <textarea class="instructional-text-area" id="instructionsText" width=>Click Start to Play!</textarea> -->
             <section class="game-area">
                 <section class="table">
                     <section class="dealer">
-
                     </section>
                     <section class="player">
                         <span class="player-chips"></span>
                         <span class="player-hand-total"></span>
                         <span class="player-current-bet"></span>
                     </section>
-                    <!-- <section class="playing-card">
-                        <section class="container">
-                            <span class="value">Q</span>
-                            <span class="suit">♥</span>
-                        </section>
-                    </section> -->
-
-                    
                 </section>
-
                 <!-- <textarea class="hand-text-area" id="dealerHand"></textarea>
                 <textarea class="hand-text-area" id="playerHand"></textarea> -->
             </section>
-
-            <form class="buttons-area">
+            <div class="buttons-area">
                 <button class="button" id="startGame">Start Game</button>
-            </form>
+            </div>
        </main>
     </div>
     <script src="main.js"></script>
